Created Event Emmiter, changed ConnectionManager to ConnectionEvents, Created event connect in ConnectionEvents
